feat(DTFS2-7224): remove all references to FF by name
