+++ .env.sample
@@ -115,7 +115,6 @@ CONTACT_US_SELF_SERVICE_PORTAL_URL=
 CHANGE_STREAM_ENABLED=false
 CLAMAV_SCANNING_ENABLED=false
 FF_TFM_PAYMENT_RECONCILIATION_ENABLED=false
-FF_TFM_FACILITY_END_DATE_ENABLED=false
 FF_TFM_DEAL_CANCELLATION_ENABLED=false
 FF_PORTAL_FACILITY_AMENDMENTS_ENABLED=false
 FF_TFM_SSO_ENABLED=false
