+++ docker-compose.yml
@@ -126,7 +126,6 @@ services:
       - CONTACT_US_EMAIL_ADDRESS
       - FF_TFM_PAYMENT_RECONCILIATION_ENABLED
       - FF_FEE_RECORD_CORRECTION_ENABLED
-      - FF_TFM_FACILITY_END_DATE_ENABLED
       - FF_TFM_DEAL_CANCELLATION_ENABLED
       - FF_TFM_SSO_ENABLED
       - ENTRA_ID_CLIENT_ID
@@ -177,7 +176,6 @@ services:
       - RATE_LIMIT_THRESHOLD
       - CHANGE_STREAM_ENABLED
       - DELETION_AUDIT_LOGS_TTL_SECONDS
-      - FF_TFM_FACILITY_END_DATE_ENABLED
       - FF_TFM_DEAL_CANCELLATION_ENABLED
       - ENTRA_ID_CLIENT_ID
       - ENTRA_ID_CLOUD_INSTANCE
