+++ portal/server/controllers/dashboard/facilities/template-filters.test.js
@@ -1,4 +1,4 @@
-import { FACILITY_TYPE } from '@ukef/dtfs2-common';
+import { FACILITY_TYPE, FACILITY_STAGE } from '@ukef/dtfs2-common';
 import { typeFilters, facilitiesTemplateFilters, createdByYouFilter, stageFilters } from './template-filters';
 import { FIELD_NAMES } from '../../../constants';
 import { DASHBOARD_FILTERS } from '../../../content-strings';
@@ -42,6 +42,10 @@ describe('controllers/dashboard/facilities - template-filters', () => {
         {
           text: DASHBOARD_FILTERS.BESPOKE_FILTER_VALUES.FACILITIES.UNISSUED,
           value: DASHBOARD_FILTERS.BESPOKE_FILTER_VALUES.FACILITIES.UNISSUED,
+        },
+        {
+          text: FACILITY_STAGE.RISK_EXPIRED,
+          value: FACILITY_STAGE.RISK_EXPIRED,
         },
       ];
