+++ trade-finance-manager-api/src/v1/rest-mappings/mappings/facilities/getLatestFacilityEndDateValuest.ff-test.ts
@@ -23,25 +23,23 @@ describe('getLatestFacilityEndDateValues', () => {
     bankReviewDate: undefined,
   };
 
-  describe('When `FF_TFM_FACILITY_END_DATE_ENABLED` is set to true', () => {
-    describe('when there are amendments', () => {
-      it('should return the the latest amendment value when the FED has been amended', () => {
-        // Arrange
-        const facilityWithAmendments: TfmFacility = {
-          ...facility,
-          facilitySnapshot: facilitySnapshotWithFacilityEndDate,
-          amendments: [
-            { ...MOCK_AMENDMENT, updatedAt: 1723653111, version: 1, tfm: { bankReviewDate: mockBankReviewDate, isUsingFacilityEndDate: false } },
-            { ...MOCK_AMENDMENT, updatedAt: 1723653222, version: 2 },
-          ],
-        };
+  describe('when there are amendments', () => {
+    it('should return the the latest amendment value when the FED has been amended', () => {
+      // Arrange
+      const facilityWithAmendments: TfmFacility = {
+        ...facility,
+        facilitySnapshot: facilitySnapshotWithFacilityEndDate,
+        amendments: [
+          { ...MOCK_AMENDMENT, updatedAt: 1723653111, version: 1, tfm: { bankReviewDate: mockBankReviewDate, isUsingFacilityEndDate: false } },
+          { ...MOCK_AMENDMENT, updatedAt: 1723653222, version: 2 },
+        ],
+      };
 
-        // Act
-        const result = getLatestFacilityEndDateValues(facilityWithAmendments);
+      // Act
+      const result = getLatestFacilityEndDateValues(facilityWithAmendments);
 
-        // Assert
-        expect(result).toEqual({ isUsingFacilityEndDate: false, bankReviewDate: mockBankReviewDate });
-      });
+      // Assert
+      expect(result).toEqual({ isUsingFacilityEndDate: false, bankReviewDate: mockBankReviewDate });
     });
   });
 });
