+++ trade-finance-manager-ui/server/controllers/helpers/deal-cancellation-enabled.helper.test.ts
@@ -15,41 +15,61 @@ const { AIN, MIN, MIA } = DEAL_SUBMISSION_TYPE;
 const { COMPLETED, PENDING, DRAFT } = TFM_DEAL_CANCELLATION_STATUS;
 
 describe('isDealCancellationEnabledForUser', () => {
-  describe('when `FF_TFM_FACILITY_END_DATE_ENABLED` is set to false', () => {
-    describe('when user is a PIM user', () => {
-      it.each([MIA, MIN, AIN])('should return false if the deal type is %s', (type) => {
-        const result = isDealCancellationEnabledForUser(type, pimUser);
+  describe('when user is a PIM user', () => {
+    it('should return false if the deal type is MIA', () => {
+      const result = isDealCancellationEnabledForUser(MIA, pimUser);
 
-        expect(result).toEqual(false);
-      });
+      expect(result).toEqual(false);
     });
 
-    describe('when user is not a PIM user', () => {
-      it.each([MIA, MIN, AIN])('should return false if the deal type is %s', (type) => {
-        const result = isDealCancellationEnabledForUser(type, nonPimUser);
+    it('should return true if the deal type is MIN', () => {
+      const result = isDealCancellationEnabledForUser(MIN, pimUser);
 
-        expect(result).toEqual(false);
-      });
+      expect(result).toEqual(true);
+    });
+
+    it('should return true if the deal type is AIN', () => {
+      const result = isDealCancellationEnabledForUser(AIN, pimUser);
+
+      expect(result).toEqual(true);
+    });
+  });
+
+  describe('when user is not a PIM user', () => {
+    it.each([MIA, MIN, AIN])('should return false if the deal type is %s', (type) => {
+      const result = isDealCancellationEnabledForUser(type, nonPimUser);
+
+      expect(result).toEqual(false);
     });
   });
 });
 
 describe('isDealCancellationEnabled', () => {
-  describe('when `FF_TFM_FACILITY_END_DATE_ENABLED` is set to false', () => {
-    describe('when user is a PIM user', () => {
-      it.each([MIA, MIN, AIN])('should return false if the deal type is %s', (type) => {
-        const result = isDealCancellationEnabled(type);
+  describe('when user is a PIM user', () => {
+    it('should return false if the deal type is MIA', () => {
+      const result = isDealCancellationEnabled(MIA);
 
-        expect(result).toEqual(false);
-      });
+      expect(result).toEqual(false);
     });
 
-    describe('when user is not a PIM user', () => {
-      it.each([MIA, MIN, AIN])('should return false if the deal type is %s', (type) => {
-        const result = isDealCancellationEnabled(type);
+    it('should return true if the deal type is MIN', () => {
+      const result = isDealCancellationEnabled(MIN);
 
-        expect(result).toEqual(false);
-      });
+      expect(result).toEqual(true);
+    });
+
+    it('should return true if the deal type is AIN', () => {
+      const result = isDealCancellationEnabled(AIN);
+
+      expect(result).toEqual(true);
+    });
+  });
+
+  describe('when user is not a PIM user', () => {
+    it.each([MIA, MIN, AIN])('should return false if the deal type is %s', (type) => {
+      const result = isDealCancellationEnabledForUser(type, nonPimUser);
+
+      expect(result).toEqual(false);
     });
   });
 });
